Update mm.c
all of malloc is done, except the add_free method (to join splitted blocks to free list).

In the definition section, I specified where the pointer of free list or allocated list should point to. 

I emailed Brandon to ask if the payload should be padded to 0. 

No replies from Brandon yet.